2014/10-16 PyCon PL Amazing DRF/code/requirements.txt
Django==1.7
djangorestframework==2.4.3

django-rest-swagger==0.1.14
gevent==1.0.1
gunicorn==19.1.1
